+++ givernylocal/src/givernylocal/turbulence_dataset.py
@@ -62,7 +62,18 @@ class turb_dataset():
         self.var = var
         self.var_offsets = var_offsets
         # convert the timepoint to [hour, minute, simulation number] for the windfarm datasets.
-        if self.dataset_title in ['diurnal_windfarm', 'nbl_windfarm']:
+        if self.dataset_title == 'diurnal_windfarm':
+            if self.var in ['force', 'pressure', 'sgsviscosity', 'soiltemperature', 'temperature', 'velocity']:
+                simulation_num = timepoint % 120
+                minute = math.floor(timepoint / 120) % 60
+                hour = math.floor((timepoint / 120) / 60)
+            elif self.var in ['heatflux', 'meanpressure', 'meantemperature', 'meanvelocity', 'reynoldsstresses', 'tempvariance']:
+                simulation_num = 0
+                minute = timepoint % 6
+                hour = math.floor(timepoint / 6)
+            
+            self.timepoint = [hour, minute, simulation_num]
+        elif self.dataset_title == 'nbl_windfarm':
             simulation_num = timepoint % 120
             minute = math.floor(timepoint / 120) % 60
             hour = math.floor((timepoint / 120) / 60)